mission17/mission17/requirements.txt
streamlit==1.49.0
streamlit-drawable-canvas==0.9.3
onnxruntime==1.22.1
pillow
numpy
pandas